pyproject.toml
[project]
name = "eap-dataset"
version = "0.1.0"
description = "eAP dataset annotation processing toolkit"
requires-python = ">=3.10.0,<3.11.0"
dependencies = [
    "opencv-python==4.11.0.86",
    "numpy==2.2.6",
    "tqdm==4.67.1",
    "PyYAML==6.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["tools"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
